% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1_filter_authorization.R
\name{authorizer}
\alias{authorizer}
\title{Authorizes incoming request (demonstration purposes only)}
\usage{
authorizer(req, res)
}
\arguments{
\item{req}{HTTP request object (plumber)}

\item{res}{HTTP response object (plumber)}
}
\description{
Authorizes incoming request (demonstration purposes only)
}
