% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1_endpoint_average_lifespan.R
\name{average_lifespan}
\alias{average_lifespan}
\title{Returns expected lifespan (UK) given age and sex}
\usage{
average_lifespan(age, sex)
}
\arguments{
\item{age}{age in years}

\item{sex}{"male" or "female"}
}
\description{
Returns expected lifespan (UK) given age and sex
}
