% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.0_endpoint_cvd_event_risk.R
\name{cvd_event_risk}
\alias{cvd_event_risk}
\title{Calculates 10-year risk of cardiovascular disease}
\usage{
cvd_event_risk(sex, age, bmi, sbp, bp_med, smoker, diabetes)
}
\arguments{
\item{sex}{"male" or "female"}

\item{age}{age in years}

\item{bmi}{body-mass index}

\item{sbp}{Systolic blood pressure}

\item{bp_med}{On blood pressure medication 1=Yes, 0=No}

\item{smoker}{Are you a smoker 1=Yes, 0=No}

\item{diabetes}{Do you have diabetes 1=Yes, 0=No}
}
\description{
Calculates 10-year risk of cardiovascular disease
}
