% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.0_helper_is_swagger_request.R
\name{is_swagger_request}
\alias{is_swagger_request}
\title{Checks if HTTP request is from Swagger}
\usage{
is_swagger_request(path_info)
}
\arguments{
\item{path_info}{Path info from a HTTP request object (plumber)}
}
\description{
Checks if HTTP request is from Swagger
}
