% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.0_filter_logging.R
\name{logger}
\alias{logger}
\title{Logs some information about the incoming HTTP request}
\usage{
logger(req)
}
\arguments{
\item{req}{HTTP request object (plumber)}
}
\description{
Logs some information about the incoming HTTP request
}
