% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.0_run_API.R
\name{run_api}
\alias{run_api}
\title{Run the API on a given host and port}
\usage{
run_api(host = "0.0.0.0", port = 80)
}
\arguments{
\item{host}{IP of the host, Default: 0.0.0.0}

\item{port}{Which port to use for the API, Default: 80}
}
\value{
A running Plumber API
}
\description{
Run the API on a given host and port
}
\examples{
\dontrun{
run_api()
run_api(host = "127.0.0.1", port = 4242)}
}
